# swagger.yml
definitions:
  Character:
    type: object
    required:
      - id
      - name
      - status
      - gender
      - image
      - url
      - created
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      status:
        type: string
      gender:
        type: string
      image:
        type: string
      url:
        type: string
      created:
        type: string
...
  ErrorMessage:
    type: object
    required:
      - message
    properties:
      message:
        type: string
paths:
...
  /readCsvFileWorkerPool:
    get:
      tags: ["readCsvWorkerPool"]
      operationId: ReadCsvWorkerPool
      parameters:
        - name: items
          in: query
          type: integer
          format: int64
          required: true
        - name: "items_per_worker"
          in: query
          type: integer
          format: int64
          required: true
        - name: "id_type"
          in: query
          type: string
          required: true
      responses:
        200:
          description: Read data from csv file using worker pool
          schema:
            type: array
            characters:
              $ref: "#/definitions/Character"
        default:
          description: generic error response
          schema:
            $ref: "#/definitions/ErrorMessage"
consumes:
  - application/json
info:
  description: 2022Q2G0 Bootcamp Final Project
  title: 2022Q2G0 Bootcamp Project
  version: 1.0.0
produces:
  - application/json
schemes:
  - http
swagger: "2.0"